+++ src/tests/parser_test.py
@@ -530,12 +530,59 @@ def test_complex_module():
                    'bar'
         ]
     """),
+    CodeSnippet("""\
+        __all__ = 'foo', 'bar'
+    """),
+    CodeSnippet("""\
+        __all__ = 'foo', 'bar',
+    """),
+    CodeSnippet(
+        """__all__ = 'foo', 'bar'"""
+    ),
+    CodeSnippet("""\
+        __all__ = 'foo', \
+                  'bar'
+    """),
+    CodeSnippet("""\
+        foo = 1
+        __all__ = 'foo', 'bar'
+    """),
+    CodeSnippet("""\
+        __all__ = 'foo', 'bar'
+        foo = 1
+    """),
+    CodeSnippet("""\
+        __all__ = ['foo', 'bar']  # never freeze
+    """),
 ))
 def test_dunder_all(code):
     """Test that __all__ is parsed correctly."""
     parser = Parser()
     module = parser.parse(code, "filepath")
     assert module.dunder_all == ('foo', 'bar')
+
+
+def test_single_value_dunder_all():
+    """Test that single value __all__ is parsed correctly."""
+    parser = Parser()
+    code = CodeSnippet("""\
+        __all__ = 'foo',
+    """)
+    module = parser.parse(code, "filepath")
+    assert module.dunder_all == ('foo', )
+
+    code = CodeSnippet("""\
+        __all__ = 'foo'
+    """)
+    module = parser.parse(code, "filepath")
+    assert module.dunder_all is None
+    assert module.dunder_all_error
+
+    code = CodeSnippet("""\
+        __all__ = ('foo', )
+    """)
+    module = parser.parse(code, "filepath")
+    assert module.dunder_all == ('foo', )
 
 
 indeterminable_dunder_all_test_cases = [