+++ src/tests/test_cases/sections.py
@@ -7,12 +7,13 @@ expect = expectation.expect
 
 
 _D213 = 'D213: Multi-line docstring summary should start at the second line'
+_D400 = "D400: First line should end with a period (not '!')"
 
 
 @expect(_D213)
 @expect("D405: Section name should be properly capitalized "
         "('Returns', not 'returns')")
-def not_capitalized():
+def not_capitalized():  # noqa: D416
     """Toggle the gizmo.
 
     returns
@@ -25,7 +26,7 @@ def not_capitalized():
 @expect(_D213)
 @expect("D406: Section name should end with a newline "
         "('Returns', not 'Returns:')")
-def superfluous_suffix():
+def superfluous_suffix():  # noqa: D416
     """Toggle the gizmo.
 
     Returns:
@@ -37,7 +38,7 @@ def superfluous_suffix():
 
 @expect(_D213)
 @expect("D407: Missing dashed underline after section ('Returns')")
-def no_underline():
+def no_underline():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -48,7 +49,8 @@ def no_underline():
 
 @expect(_D213)
 @expect("D407: Missing dashed underline after section ('Returns')")
-def no_underline_and_no_description():
+@expect("D414: Section has no content ('Returns')")
+def no_underline_and_no_description():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -61,7 +63,7 @@ def no_underline_and_no_description():
 @expect("D414: Section has no content ('Returns')")
 @expect("D411: Missing blank line before section ('Yields')")
 @expect("D414: Section has no content ('Yields')")
-def consecutive_sections():
+def consecutive_sections():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -79,7 +81,7 @@ def consecutive_sections():
 @expect(_D213)
 @expect("D408: Section underline should be in the line following the "
         "section's name ('Returns')")
-def blank_line_before_underline():
+def blank_line_before_underline():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -93,7 +95,7 @@ def blank_line_before_underline():
 @expect(_D213)
 @expect("D409: Section underline should match the length of its name "
         "(Expected 7 dashes in section 'Returns', got 2)")
-def bad_underline_length():
+def bad_underline_length():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -105,7 +107,7 @@ def bad_underline_length():
 
 @expect(_D213)
 @expect("D413: Missing blank line after last section ('Returns')")
-def no_blank_line_after_last_section():
+def no_blank_line_after_last_section():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -116,7 +118,7 @@ def no_blank_line_after_last_section():
 
 @expect(_D213)
 @expect("D411: Missing blank line before section ('Returns')")
-def no_blank_line_before_section():
+def no_blank_line_before_section():  # noqa: D416
     """Toggle the gizmo.
 
     The function's description.
@@ -129,7 +131,7 @@ def no_blank_line_before_section():
 
 @expect(_D213)
 @expect("D214: Section is over-indented ('Returns')")
-def section_overindented():
+def section_overindented():  # noqa: D416
     """Toggle the gizmo.
 
         Returns
@@ -141,7 +143,7 @@ def section_overindented():
 
 @expect(_D213)
 @expect("D215: Section underline is over-indented (in section 'Returns')")
-def section_underline_overindented():
+def section_underline_overindented():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -155,7 +157,7 @@ def section_underline_overindented():
 @expect("D215: Section underline is over-indented (in section 'Returns')")
 @expect("D413: Missing blank line after last section ('Returns')")
 @expect("D414: Section has no content ('Returns')")
-def section_underline_overindented_and_contentless():
+def section_underline_overindented_and_contentless():  # noqa: D416
     """Toggle the gizmo.
 
     Returns
@@ -164,7 +166,7 @@ def section_underline_overindented_and_contentless():
 
 
 @expect(_D213)
-def ignore_non_actual_section():
+def ignore_non_actual_section():  # noqa: D416
     """Toggle the gizmo.
 
     This is the function's description, which will also specify what it
@@ -175,11 +177,13 @@ def ignore_non_actual_section():
 
 @expect(_D213)
 @expect("D401: First line should be in imperative mood "
-        "('Return', not 'Returns')")
+        "(perhaps 'Return', not 'Returns')")
 @expect("D400: First line should end with a period (not 's')")
+@expect("D415: First line should end with a period, question "
+        "mark, or exclamation point (not 's')")
 @expect("D205: 1 blank line required between summary line and description "
         "(found 0)")
-def section_name_in_first_line():
+def section_name_in_first_line():  # noqa: D416
     """Returns
     -------
     A value of some sort.
@@ -199,7 +203,7 @@ def section_name_in_first_line():
 @expect("D406: Section name should end with a newline "
         "('Raises', not 'Raises:')")
 @expect("D407: Missing dashed underline after section ('Raises')")
-def multiple_sections():
+def multiple_sections():  # noqa: D416
     """Toggle the gizmo.
 
     Short summary
@@ -218,7 +222,7 @@ def multiple_sections():
 
 
 @expect(_D213)
-def false_positive_section_prefix():
+def false_positive_section_prefix():  # noqa: D416
     """Toggle the gizmo.
 
     Parameters
@@ -229,7 +233,7 @@ def false_positive_section_prefix():
 
 
 @expect(_D213)
-def section_names_as_parameter_names():
+def section_names_as_parameter_names():  # noqa: D416
     """Toggle the gizmo.
 
     Parameters
@@ -238,5 +242,46 @@ def section_names_as_parameter_names():
         A list of wonderful notes.
     examples: list
         A list of horrible examples.
+
+    """
+
+
+@expect(_D213)
+@expect("D414: Section has no content ('Returns')")
+def valid_google_style_section():  # noqa: D406, D407
+    """Toggle the gizmo.
+
+    Args:
+        note: A random string.
+
+    Returns:
+
+    Raises:
+        RandomError: A random error that occurs randomly.
+
+    """
+
+
+@expect(_D213)
+@expect("D416: Section name should end with a semicolon "
+        "('Args:', not 'Args')")
+def missing_colon_google_style_section():  # noqa: D406, D407
+    """Toggle the gizmo.
+
+    Args
+        note: A random string.
+
+    """
+
+
+@expect(_D213)
+@expect("D417: Missing arguments in the function docstring "
+        "(argument(s) 'y' missing in function "
+        "'test_missing_args' docstring)")
+def test_missing_args(x=1, y=2):  # noqa: D407, D407
+    """Toggle the gizmo.
+
+    Args:
+        x (int): The greatest integer.
 
     """